refactor(types): Change a way to define types to use in JSDoc notations
