+++ lib/connections/mongoose.js
@@ -3,7 +3,7 @@ import Promise from 'bluebird';
 import logger from 'lib/logger';
 import defaultTo from 'lodash/defaultTo';
 
-/** @typedef {module:mongoose.Connection} Connection */
+/** @typedef {import("./mongoose.types").Connection} Connection */
 
 mongoose.Promise = Promise;
 mongoose.set('useNewUrlParser', true);
