+++ lib/connections/mongoose.types.ts
@@ -1,0 +1,3 @@
+import { Connection } from 'mongoose';
+
+export { Connection };
