+++ lib/models/personasImport.js
@@ -34,12 +34,12 @@ import {
 } from 'lib/constants/personasImport';
 import moment from 'moment';
 
-/** @typedef {module:lib.models.personasImport.PersonasImportStructure} PersonasImportStructure */
-/** @typedef {module:lib.models.personasImport.PersonaImportField} PersonaImportField */
-/** @typedef {module:lib.models.personasImport.PersonasImport} PersonasImport */
-/** @typedef {module:lib.models.personasImport.PersonasImportError} PersonasImportError */
-/** @typedef {module:lib.models.personasImport.PersonasImportResult} PersonasImportResult */
-/** @typedef {module:lib.models.personasImport.PersonasImportSchema} PersonasImportSchema */
+/** @typedef {import("./personasImport.types").PersonasImportStructure} PersonasImportStructure */
+/** @typedef {import("./personasImport.types").PersonaImportField} PersonaImportField */
+/** @typedef {import("./personasImport.types").PersonasImport} PersonasImport */
+/** @typedef {import("./personasImport.types").PersonasImportError} PersonasImportError */
+/** @typedef {import("./personasImport.types").PersonasImportResult} PersonasImportResult */
+/** @typedef {import("./personasImport.types").PersonasImportSchema} PersonasImportSchema */
 
 /**
  * @param {PersonasImportStructure} structure
